refactored out AdminPage components
