+++ __tests__/AdminPage/AdminPage.test.js
@@ -3,7 +3,7 @@
  */
 import React from 'react';
 import { render, screen } from '../../test-utils';
-import AdminPage from '../../components/Admin/AdminPage';
+import AdminPage from '../../components/AdminPage/AdminPage';
 import InitialState from '../../store/Store/InitialState';
 import { waitFor } from '@testing-library/react';
 import Constants from '../../test-utils/constant';
