+++ components/AdminPage/AdminModal.js
@@ -14,7 +14,7 @@ const AdminModal = ({ setModal, modal, bounty, payoutTokenAddress }) => {
   const [appState] = useContext(StoreContext);
 
   useEffect(() => {
-    if (modal.transaction && (modal.type === 'Budget' || modal.type === 'Payout')) {
+    if (modal && modal.transaction && (modal.type === 'Budget' || modal.type === 'Payout')) {
       const tokenAddress = modal.transaction.events[0].args[1];
       const token = appState.tokenClient.getToken(tokenAddress);
       setToken(token);
@@ -23,10 +23,6 @@ const AdminModal = ({ setModal, modal, bounty, payoutTokenAddress }) => {
       let decimals = parseInt(token.decimals) || 18;
       let formattedVolume = ethers.utils.formatUnits(bigNumberVolume, decimals);
       setVolume(formattedVolume);
-    }
-
-    if (modal.transaction && modal.type === 'PayoutSchedule') {
-      // do something?
     }
   }, [modal]);
 
@@ -59,7 +55,7 @@ const AdminModal = ({ setModal, modal, bounty, payoutTokenAddress }) => {
     // Courtesy of https://stackoverflow.com/questions/32553158/detect-click-outside-react-component
     function handleClickOutside(event) {
       if (modalRef.current && !modalRef.current.contains(event.target)) {
-        setModal();
+        // setModal({});
       }
     }
 
@@ -73,7 +69,7 @@ const AdminModal = ({ setModal, modal, bounty, payoutTokenAddress }) => {
   }, [modal, modalRef]);
 
   const closeModal = () => {
-    setModal(false);
+    // setModal({});
   };
 
   const title = {
@@ -103,7 +99,7 @@ const AdminModal = ({ setModal, modal, bounty, payoutTokenAddress }) => {
   };
 
   const btn =
-    modal.type.includes('Closed') || modal.type === 'Error' || modal.type === 'Loading' ? (
+    modal.type?.includes('Closed') || modal.type === 'Error' || modal.type === 'Loading' ? (
       <button onClick={closeModal} className='btn-default'>
         <span>Close</span>
       </button>
@@ -112,93 +108,34 @@ const AdminModal = ({ setModal, modal, bounty, payoutTokenAddress }) => {
     );
 
   return (
-    <div ref={modalRef}>
-      <ModalDefault title={title[modal.type]} footerRight={btn} setShowModal={setModal} resetState={setModal}>
-        {(modal.type === 'Payout' || modal.type === 'Budget') && token && (
-          <>
-            <div className='gap-4 grid grid-cols-[150px_1fr]'>
-              <div className='flex whitespace-pre-wrap col-span-2'>{content[modal.type]}</div>
-              <div>{modal.type} set to:</div>
-              <div className='flex gap-2'>
-                <Image
-                  alt='crypto-logo'
-                  width={24}
-                  className='inline'
-                  height={24}
-                  src={token.path || token.logoURI || '/crypto-logos/ERC20.svg'}
-                />
-                <span>
-                  {volume} {token.symbol}
-                </span>
-              </div>
-              <div>Value: </div>
-              <div>{appState.utils.formatter.format(modal.type === 'Budget' ? budget : split)}</div>
-              <div className='flex-1' href={modal.transaction.transactionHash}>
-                Transaction:
-              </div>
-              <a
-                className='break-all flex-1 underline cursor-pointer'
-                target='_blank'
-                rel='noopener noreferrer'
-                href={`${process.env.NEXT_PUBLIC_BLOCK_EXPLORER_BASE_URL}/tx/${modal.transaction.transactionHash}`}
-              >
-                {modal.transaction.transactionHash.slice(0, 3)}...
-                {modal.transaction.transactionHash.slice(-3)}
-                <LinkText />
-              </a>
-            </div>
-          </>
-        )}
-
-        {modal.type === 'PayoutSchedule' && (
-          <>
-            <div className='gap-4 grid grid-cols-[200px_1fr]'>
-              <div className='flex whitespace-pre-wrap col-span-2'>{content[modal.type]}</div>
-              <div>Payout Schedule set to:</div>
-              <div className='grid grid-cols-[120px_1fr] text-xs font-semibold leading-loose'>
-                <div>Number of tiers:</div>
-                <div>{modal.finalTierVolume.length}</div>
-
-                {modal.finalTierVolume.map((t, index) => {
-                  return (
-                    <div key={index} className='col-span-2 grid grid-cols-[120px_1fr]'>
-                      <div>{`${appState.utils.handleSuffix(index + 1)} winner:`}</div>
-                      <div className='self-center'>
-                        {t} {getVolumeSuffix(bounty.bountyType, payoutTokenAddress)}
-                      </div>
-                    </div>
-                  );
-                })}
-              </div>
-              <div href={modal.transaction.transactionHash}>Transaction:</div>
-              <a
-                className='break-all flex-1 underline cursor-pointer'
-                target='_blank'
-                rel='noopener noreferrer'
-                href={`${process.env.NEXT_PUBLIC_BLOCK_EXPLORER_BASE_URL}/tx/${modal.transaction.transactionHash}`}
-              >
-                {modal.transaction.transactionHash.slice(0, 3)}...
-                {modal.transaction.transactionHash.slice(-3)}
-                <LinkText />
-              </a>
-            </div>
-          </>
-        )}
-
-        {modal.type === 'Loading' && (
-          <>
-            <div className='flex items-center gap-2'>
-              Your request is being processed... <LoadingIcon />
-            </div>
-          </>
-        )}
-
-        {(modal.type.includes('Closed') || modal.type === 'Error') && (
-          <>
-            <div className='gap-4 grid grid-cols-[150px_1fr]'>
-              <div className='pb-2 whitespace-pre-wrap col-span-2'>{content[modal.type]}</div>
-              {modal.type !== 'Error' && (
-                <>
+    <>
+      {modal.type && (
+        <div ref={modalRef}>
+          <ModalDefault
+            title={title[modal.type]}
+            footerRight={btn}
+            setShowModal={() => setModal({})}
+            resetState={() => setModal({})}
+          >
+            {(modal.type === 'Payout' || modal.type === 'Budget') && token && (
+              <>
+                <div className='gap-4 grid grid-cols-[150px_1fr]'>
+                  <div className='flex whitespace-pre-wrap col-span-2'>{content[modal.type]}</div>
+                  <div>{modal.type} set to:</div>
+                  <div className='flex gap-2'>
+                    <Image
+                      alt='crypto-logo'
+                      width={24}
+                      className='inline'
+                      height={24}
+                      src={token.path || token.logoURI || '/crypto-logos/ERC20.svg'}
+                    />
+                    <span>
+                      {volume} {token.symbol}
+                    </span>
+                  </div>
+                  <div>Value: </div>
+                  <div>{appState.utils.formatter.format(modal.type === 'Budget' ? budget : split)}</div>
                   <div className='flex-1' href={modal.transaction.transactionHash}>
                     Transaction:
                   </div>
@@ -212,13 +149,81 @@ const AdminModal = ({ setModal, modal, bounty, payoutTokenAddress }) => {
                     {modal.transaction.transactionHash.slice(-3)}
                     <LinkText />
                   </a>
-                </>
-              )}
-            </div>
-          </>
-        )}
-      </ModalDefault>
-    </div>
+                </div>
+              </>
+            )}
+
+            {modal.type === 'PayoutSchedule' && (
+              <>
+                <div className='gap-4 grid grid-cols-[200px_1fr]'>
+                  <div className='flex whitespace-pre-wrap col-span-2'>{content[modal.type]}</div>
+                  <div>Payout Schedule set to:</div>
+                  <div className='grid grid-cols-[120px_1fr] text-xs font-semibold leading-loose'>
+                    <div>Number of tiers:</div>
+                    <div>{modal.finalTierVolume.length}</div>
+
+                    {modal.finalTierVolume.map((t, index) => {
+                      return (
+                        <div key={index} className='col-span-2 grid grid-cols-[120px_1fr]'>
+                          <div>{`${appState.utils.handleSuffix(index + 1)} winner:`}</div>
+                          <div className='self-center'>
+                            {t} {getVolumeSuffix(bounty.bountyType, payoutTokenAddress)}
+                          </div>
+                        </div>
+                      );
+                    })}
+                  </div>
+                  <div href={modal.transaction.transactionHash}>Transaction:</div>
+                  <a
+                    className='break-all flex-1 underline cursor-pointer'
+                    target='_blank'
+                    rel='noopener noreferrer'
+                    href={`${process.env.NEXT_PUBLIC_BLOCK_EXPLORER_BASE_URL}/tx/${modal.transaction.transactionHash}`}
+                  >
+                    {modal.transaction.transactionHash.slice(0, 3)}...
+                    {modal.transaction.transactionHash.slice(-3)}
+                    <LinkText />
+                  </a>
+                </div>
+              </>
+            )}
+
+            {modal.type === 'Loading' && (
+              <>
+                <div className='flex items-center gap-2'>
+                  Your request is being processed... <LoadingIcon />
+                </div>
+              </>
+            )}
+
+            {(modal.type.includes('Closed') || modal.type === 'Error') && (
+              <>
+                <div className='gap-4 grid grid-cols-[150px_1fr]'>
+                  <div className='pb-2 whitespace-pre-wrap col-span-2'>{content[modal.type]}</div>
+                  {modal.type !== 'Error' && (
+                    <>
+                      <div className='flex-1' href={modal.transaction.transactionHash}>
+                        Transaction:
+                      </div>
+                      <a
+                        className='break-all flex-1 underline cursor-pointer'
+                        target='_blank'
+                        rel='noopener noreferrer'
+                        href={`${process.env.NEXT_PUBLIC_BLOCK_EXPLORER_BASE_URL}/tx/${modal.transaction.transactionHash}`}
+                      >
+                        {modal.transaction.transactionHash.slice(0, 3)}...
+                        {modal.transaction.transactionHash.slice(-3)}
+                        <LinkText />
+                      </a>
+                    </>
+                  )}
+                </div>
+              </>
+            )}
+          </ModalDefault>
+        </div>
+      )}
+    </>
   );
 };
